LayoutTests/imported/w3c/web-platform-tests/css/css-tables/visibility-collapse-rowcol-002-expected.txt

Visibility collapse

Spec
Setting row and spanning column to visibility:collapse changes height and width.


FAIL spanning row visibility:collapse doesn't change table width assert_equals: spanning row visibility:collapse doesn't change table width expected 45 but got 130
FAIL spanning row visibility:collapse doesn't change height in this case assert_equals: spanning row visibility:collapse doesn't change height in this case expected 35 but got 75
